hotfix/be/#160 event 엔티티, dto common으로 이동
common으로 이동 및 엔티티에서 Admin -> Long 형변환으로 인한 오류 해결
