+++ admin/src/main/java/kernel/maidlab/admin/event/repository/EventRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import kernel.maidlab.admin.event.entity.Event;
+import kernel.maidlab.common.entity.event.Event;
 
 @Repository
 public interface EventRepository extends JpaRepository<Event, Long> {
